+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+[0.6.0]
+- Added: `GoFuncWrapper` wrapper function that allows extending BRE with strongly typed functions
+- Changed: updated README to reflect the improvement
+
 [0.5.0]
 - Added: custom reusable js functions maybe defined in `scripts` section of a rules yaml file
 - Changed: injectable debug function arguments from `(context, interface{})` to `(...interface{})`